feat: improve AsyncChunkSerializer (task grouping)
