+++ orebfuscator-plugin/src/main/java/net/imprex/orebfuscator/cache/ChunkSerializer.java
@@ -1,101 +1,92 @@
 package net.imprex.orebfuscator.cache;
 
+import java.io.DataInputStream;
+import java.io.DataOutputStream;
 import java.io.IOException;
-import java.util.HashMap;
-import java.util.LinkedList;
-import java.util.Map;
-import java.util.Queue;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.locks.Condition;
-import java.util.concurrent.locks.Lock;
-import java.util.concurrent.locks.ReentrantLock;
+import java.util.Collection;
 
+import net.imprex.orebfuscator.NmsInstance;
 import net.imprex.orebfuscator.obfuscation.ObfuscatedChunk;
+import net.imprex.orebfuscator.util.BlockCoords;
 import net.imprex.orebfuscator.util.ChunkPosition;
 
-public class ChunkSerializer implements Runnable {
+public class ChunkSerializer {
 
-	private final Lock lock = new ReentrantLock();
-	private final Condition condition = lock.newCondition();
+	private static final int CACHE_VERSION = 1;
 
-	private final Map<ChunkPosition, Runnable> tasks = new HashMap<>();
-	private final Queue<ChunkPosition> positions = new LinkedList<>();
-
-	private final AtomicBoolean suspended = new AtomicBoolean();
-	private final Thread thread;
-
-	private volatile boolean running = true;
-
-	public ChunkSerializer() {
-		this.thread = new Thread(this, "chunk-serializer");
-		this.thread.setDaemon(true);
-		this.thread.start();
+	private static DataInputStream createInputStream(ChunkPosition key) throws IOException {
+		return NmsInstance.getRegionFileCache().createInputStream(key);
 	}
 
-	public CompletableFuture<ObfuscatedChunk> read(ChunkPosition position) {
-		CompletableFuture<ObfuscatedChunk> future = new CompletableFuture<>();
-		this.queueTask(position, () -> {
-			try {
-				future.complete(ChunkCacheSerializer.read(position));
-			} catch (IOException e) {
-				e.printStackTrace();
-				future.complete(null);
-			}
-		});
-		return future;
+	private static DataOutputStream createOutputStream(ChunkPosition key) throws IOException {
+		return NmsInstance.getRegionFileCache().createOutputStream(key);
 	}
 
-	public void write(ChunkPosition position, ObfuscatedChunk chunk) {
-		this.queueTask(position, () -> {
-			try {
-				ChunkCacheSerializer.write(position, chunk);
-			} catch (IOException e) {
-				e.printStackTrace();
-			}
-		});
-	}
-
-	private void queueTask(ChunkPosition position, Runnable runnable) {
-		this.lock.lock();
-		try {
-			if (this.tasks.put(position, runnable) == null) {
-				this.positions.offer(position);
-			}
-
-			if (this.suspended.compareAndSet(true, false)) {
-				this.condition.signal();
-			}
-		} finally {
-			this.lock.unlock();
-		}
-	}
-
-	@Override
-	public void run() {
-		while (this.running || !this.positions.isEmpty()) {
-			this.lock.lock();
-			try {
-				if (this.positions.isEmpty() && this.suspended.compareAndSet(false, true)) {
-					this.condition.awaitUninterruptibly();
-				} else {
-					this.tasks.remove(this.positions.poll()).run();
+	public static ObfuscatedChunk read(ChunkPosition key) throws IOException {
+		try (DataInputStream dataInputStream = createInputStream(key)) {
+			if (dataInputStream != null) {
+				// check if cache entry has right version and if chunk is present
+				if (dataInputStream.readInt() != CACHE_VERSION || !dataInputStream.readBoolean()) {
+					return null;
 				}
-			} finally {
-				this.lock.unlock();
+
+				byte[] hash = new byte[dataInputStream.readInt()];
+				dataInputStream.readFully(hash);
+
+				byte[] data = new byte[dataInputStream.readInt()];
+				dataInputStream.readFully(data);
+
+				ObfuscatedChunk chunkCacheEntry = new ObfuscatedChunk(hash, data);
+
+				Collection<BlockCoords> proximityBlocks = chunkCacheEntry.getProximityBlocks();
+				for (int i = dataInputStream.readInt(); i > 0; i--) {
+					proximityBlocks.add(BlockCoords.fromLong(dataInputStream.readLong()));
+				}
+
+				Collection<BlockCoords> removedEntities = chunkCacheEntry.getRemovedTileEntities();
+				for (int i = dataInputStream.readInt(); i > 0; i--) {
+					removedEntities.add(BlockCoords.fromLong(dataInputStream.readLong()));
+				}
+
+				return chunkCacheEntry;
+			}
+		} catch (IOException e) {
+			throw new IOException("Unable to read chunk: " + key, e);
+		}
+		return null;
+	}
+
+	// TODO consider size limit for cache since RegionFile before 1.14 have a hard limit of 256 * 4kb 
+	public static void write(ChunkPosition key, ObfuscatedChunk value) throws IOException {
+		try (DataOutputStream dataOutputStream = createOutputStream(key)) {
+			dataOutputStream.writeInt(CACHE_VERSION);
+
+			if (value != null) {
+				dataOutputStream.writeBoolean(true);
+
+				byte[] hash = value.getHash();
+				dataOutputStream.writeInt(hash.length);
+				dataOutputStream.write(hash, 0, hash.length);
+
+				byte[] data = value.getData();
+				dataOutputStream.writeInt(data.length);
+				dataOutputStream.write(data, 0, data.length);
+
+				Collection<BlockCoords> proximityBlocks = value.getProximityBlocks();
+				dataOutputStream.writeInt(proximityBlocks.size());
+				for (BlockCoords blockPosition : proximityBlocks) {
+					dataOutputStream.writeLong(blockPosition.toLong());
+				}
+
+				Collection<BlockCoords> removedEntities = value.getRemovedTileEntities();
+				dataOutputStream.writeInt(removedEntities.size());
+				for (BlockCoords blockPosition : removedEntities) {
+					dataOutputStream.writeLong(blockPosition.toLong());
+				}	
+			} else {
+				dataOutputStream.writeBoolean(false);
 			}
 		}
 	}
 
-	public void close() {
-		this.lock.lock();
-		try {
-			this.running = false;
-			if (this.suspended.compareAndSet(true, false)) {
-				this.condition.signal();
-			}
-		} finally {
-			this.lock.unlock();
-		}
-	}
 }
